requirements.txt
django==2.1
lxml==4.1.1
pyquery==1.4.0
psycopg2-binary==2.7.5
requests[socks]==2.18.4
kafka-python==1.4.1
pyexecjs==1.5.1
apscheduler==3.5.1

djangorestframework==3.9.0
markdown==2.1.0
MySQL-python